backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
flask-cors==4.0.1
flasgger==0.9.7.1
